% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_RLumCarlo.R
\name{calc_RLumCarlo}
\alias{calc_RLumCarlo}
\title{Plot results from Monte-Carlo simulations with RLumCarlo}
\usage{
calc_RLumCarlo(results)
}
\arguments{
\item{results}{\code{\link{array}}:}
}
\value{
This function returns a \code{\link{data.frame}}
}
\description{
Plot results from Monte-Carlo simulations with RLumCarlo
}
\section{Function version}{
 0.0.1 [2017-01-27]
}

\author{
Johannes Friedrich, University of Bayreuth (Germany)
}
