% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RLumCarlo.R
\name{plot_RLumCarlo}
\alias{plot_RLumCarlo}
\title{Plot RLumCarlo Monte-Carlo Simulations Results}
\usage{
plot_RLumCarlo(object, plot_uncertainty = "range", norm = FALSE,
  add = FALSE, ...)
}
\arguments{
\item{object}{\link{list} of class \code{RLumCarlo_Model_Output} (\strong{required}): input object to be plotted,
usually the required input object is generated by one for the functions preceeding with \code{run}.
Alternatively a list of such objects can be provided.}

\item{plot_uncertainty}{\link{logical} (\emph{with default}): sets the nature of the show uncertainty, allowed
values are \code{range}, \code{sd} (standard deviation), \code{var} (variance) and \code{NULL} disables the uncertaintiy
visualisation}

\item{norm}{\link{logical} (\emph{with default}): normalise curve to the highest intensity}

\item{add}{\link{logical} (\emph{with default}): allows overplotting of results by adding curve to
an existing plot}

\item{\dots}{further arguments that can be passed to control the plot output. Currently supported
are: \code{xlab}, \code{xlim}, \code{ylim}, \code{main}, \code{lwd}, \code{type}, \code{pch}, \code{lty},\code{col}, \code{grid}, \code{legend}.
The arguments \code{lwd}, \code{type}, \code{pch}, \code{lty}, \code{col} can be provided as vector if \code{object} is a \link{list}}
}
\value{
This function returns a graphical output
}
\description{
The function allows to visualise 'RLumCarlo' modelling results without
extracting the values manually. Typically visualised values are the signal or the number of remaining electrons
as averaged values with a polygon indicating modelling uncertainties.
}
\details{
The colour curves, the package \link[khroma:khroma-package]{khroma::khroma-package} is used to provide colours
that can be best distinguished, in particular by colour-blind users.
}
\section{Function version}{
 0.1.0
}

\author{
Johannes Friedrich, University of Bayreuth (Germany), Sebastian Kreutzer, IRAMAT-CRP2A, Université
Bordeaux Montaigne (France)
} 

\section{How to cite}{
Friedrich, J., Kreutzer, S., 2019. plot_RLumCarlo(): Plot RLumCarlo Monte-Carlo Simulations Results. Function version 0.1.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-106. 
}

