% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLumCarlo-package.R
\docType{package}
\name{RLumCarlo-package}
\alias{RLumCarlo-package}
\title{\tabular{ll}{ Package: \tab RLumCarlo\cr Type: \tab Package\cr Version:
\tab 0.0.1\cr Date: \tab 2017-01-31 \cr License: \tab GPL-3\cr }}
\description{
\tabular{ll}{ Package: \tab RLumCarlo\cr Type: \tab Package\cr Version:
\tab 0.0.1\cr Date: \tab 2017-01-31 \cr License: \tab GPL-3\cr }
}
\author{
\bold{Authors}
}
