% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_CW_IRSL_LOC.R
\name{run_MC_CW_IRSL_LOC}
\alias{run_MC_CW_IRSL_LOC}
\title{Run Monte-Carlo simulation for CW-IRSL for localized transition}
\usage{
run_MC_CW_IRSL_LOC(A, times, clusters = 10, n_filled = 100, r,
  method = "par", output = "signal", ...)
}
\arguments{
\item{A}{\link{numeric} (\strong{required}): The optical excitation rate from trap to the excited state (s^-1).}

\item{times}{\link{numeric} (\emph{with default}): The sequence of time steps within the simulation (s).}

\item{clusters}{\link{numeric} (\emph{with default}): The number of MC runs (unitless).}

\item{n_filled}{\link{integer} (\emph{with default}): The number of filled electron traps at the beginning of the simulation (unitless).}

\item{r}{\link{numeric} (\emph{with default}): The localized retrapping ratio (unitless).}

\item{method}{\link{character} (\emph{with default}): sequential \code{'seq'} or parallel \code{'par'}processing}

\item{output}{\link{character} (\emph{with default}): output is either the \code{'signal'} (the default) or \code{'remaining_e'} (the remaining
charges, electrons, in the trap)}

\item{\dots}{further arguments}
}
\value{
This function returns an \link{array} with dimension length(times) x length(r) x clusters
}
\description{
Runs a Monte-Carlo (MC) simulation of constant wave infrared stimulated luminesence (CW-IRSL) using the generalized one trap (GOT) model. Localized refers to an excited state that is shared by the electron and the recombination center, so that the conduction band is not involved in the recombination process.
}
\details{
\deqn{
I_{LOC}(t) = -dn/dt = A * (n^2 / (r + n))
}

Where in the function: \cr
t := Time \cr
n := `n_filled``
}
\section{Function version}{
 0.1.0
}

\examples{
##============================================================================##
## Example 1: Single Plot for Monte-Carlo (MC) simulations for localized CW_IRSL
##============================================================================##
\dontrun{
run_MC_CW_IRSL_LOC(
 A = 0.12,
 times = 0:100,
 clusters = 50,
 n_filled = 1,
 r = 1e-7,
 method = "seq",
 output = "signal"
) \%>\%
 #Plot results of the MC simulation
plot_RLumCarlo(legend = T)

}


} 

\section{How to cite}{
Kreutzer, S., 2019. run_MC_CW_IRSL_LOC(): Run Monte-Carlo simulation for CW-IRSL for localized transition. Function version 0.1.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-102. 
}

\references{
Pagonis, V., Friedrich, J., Discher, M., Müller-Kirschbaum, A., Schlosser, V., Kreutzer, S., Chen, R. and Schmidt, C., 2019. Excited state luminescence signals from a random distribution of defects: A new Monte Carlo simulation approach for feldspar. Journal of Luminescence 207, 266–272. \doi{10.1016/j.jlumin.2018.11.024}

Reuven, C. and S. Mckeever, 1997. Theory of thermoluminescence and related phenomena.
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
}
