% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_CW_OSL_DELOC.R
\name{run_MC_CW_OSL_DELOC}
\alias{run_MC_CW_OSL_DELOC}
\title{Run Monte-Carlo simulation for CW-OSL for delocalized transition}
\usage{
run_MC_CW_OSL_DELOC(A, times, clusters = 10, N_e = 200,
  n_filled = N_e, R, method = "par", output = "signal", ...)
}
\arguments{
\item{A}{\link{numeric} (\strong{required}): The optical excitation rate from trap to conduction band (s^-1).}

\item{times}{\link{numeric} (\emph{with default}): The sequence of temperature steps within the simulation (s)}

\item{clusters}{\link{numeric} (\emph{with default}): The number of MC runs (unitless).}

\item{N_e}{\link{integer} (\emph{with default}): The total number of electron traps available (unitless).}

\item{n_filled}{\link{integer} (\emph{with default}): The number of filled electron traps at the beginning of the simulation (unitless).}

\item{R}{\link{numeric} (\emph{with default}): The delocalized retrapping ratio (unitless).}

\item{method}{\link{character} (\emph{with default}): sequential \code{'seq'} or parallel processing \code{'par'}}

\item{output}{\link{character} (\emph{with default}): output is either the \code{'signal'} (the default) or \code{'remaining_e'} (the remaining
charges, electrons, in the trap)}

\item{\dots}{further arguments}
}
\value{
This function returns an \link{array} with dimension length(times) x length(r) x clusters
}
\description{
Runs a Monte-Carlo (MC) simulation of constant wave optically stimulated luminesence (CW-OSL) using the one trap one recombination center (OTOR) model. Delocalized refers to involvement of the conduction band.
}
\details{
\deqn{
I_{DELOC}(t) = -dn/dt = p(t) * (n^2 / (N*R + n(1-R)))
}

Where in the function: \cr
\code{t} := \code{Time} \cr
\code{p(t)} := The experimental stimulation mode \cr
\code{n} := The Instantaneous number of electrons \cr
\code{N} = \code{N_e}
}
\section{Function version}{
 0.1.0
}

\examples{
##============================================================================##
## Example 1: Single Plot for Monte-Carlo (MC) simulations for delocalized CW-OSL
##============================================================================##
\dontrun{
run_MC_CW_OSL_DELOC(
 A = 0.12,
 R = 0.1,
 times = 0:100
) \%>\%
 #Plot results of the MC simulation
plot_RLumCarlo(legend = T)

}

##============================================================================##
## Example 2: Simulate CW-OSL DELOC with several parameter changes
##============================================================================##
\dontrun{

# define your parameters
A <- c(0.1,0.3,0.5,1)
times <- seq(0,60,1)
s <- 1e12
E<- 1
R<-c(1e-7,1e-6,0.01,0.1) # sequence of different R values
clusters <- 1000 # number of Monte Carlo simulations
N_e <- c(200, 500, 700, 400) # number of free electrons
n_filled <- c(200, 500, 100, 70) # number of filled traps
method <-"par"
output <- "signal"
col <- c(1,2,3,4) # ifferent colours for the individual curves
plot_uncertainty <- c(T,F,T,F)  # do you want to see the uncertainty?
add_TF <- c(F,rep(T, (length(R)-1)))
for (u in 1:length(R)){
 results <-run_MC_CW_OSL_DELOC(A=A[u], times, clusters =clusters, N_e = N_e[u],
                                n_filled = n_filled[u], R=R[u], method = method, output = output)
 plot_RLumCarlo(results,add=add_TF[u],legend = F, col=col[u], main=" your plot")
}
legend("topright",ncol=4,cex=0.55,title = "parameters" ,legend=c(paste0("A = ", A),
                                                                paste0("n_filled = ", n_filled),
                                                                paste0("N_e = ", N_e),
                                                                paste0("R = ", R)),  text.col=col)

}


} 

\section{How to cite}{
Kreutzer, S., 2019. run_MC_CW_OSL_DELOC(): Run Monte-Carlo simulation for CW-OSL for delocalized transition. Function version 0.1.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-106. 
}

\references{
Pagonis, V., Friedrich, J., Discher, M., Müller-Kirschbaum, A., Schlosser, V., Kreutzer, S., Chen, R. and Schmidt, C., 2019. Excited state luminescence signals from a random distribution of defects: A new Monte Carlo simulation approach for feldspar. Journal of Luminescence 207, 266–272. \doi{10.1016/j.jlumin.2018.11.024}

Reuven, C. and S. Mckeever, 1997. Theory of thermoluminescence and related phenomena.
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
}
