% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_ISO.R
\name{run_MC_ISO}
\alias{run_MC_ISO}
\title{Run Monte-Carlo simulation for isothermal measurements}
\usage{
run_MC_ISO(A, rho, times, clusters = 10, r = NULL, N_e = 200,
  method = "par", output = "signal", ...)
}
\arguments{
\item{A}{\code{\link{numeric}}}

\item{rho}{\code{\link{numeric}}}

\item{times}{\code{\link{vector}} (with default)}

\item{clusters}{\code{\link{numeric}} (with default):}

\item{r}{\code{\link{numeric}} (with default)}

\item{N_e}{\code{\link{numeric}} (with default):}

\item{method}{\code{\link{character}} (with default):}

\item{output}{\code{\link{character}} (with default):}

\item{\dots}{further arguments}
}
\value{
This function returns a list.
}
\description{
Run Monte-Carlo simulation for isothermal measurements
}
\section{Function version}{
 0.0.1 [2017-01-27]
}

\examples{
\dontrun{
##============================================================================##
## Example 1: Simulate isothermal measurement
##============================================================================##

times <- seq(0, 500)
run_MC_ISO(A = 0.20,
          rho = 0.007,
          times = times) \%>\%
  calc_RLumCarlo() \%>\%
    plot_RLumCarlo(legend = T)
}
}
\references{
Pagonis 2017
}
\author{
Johannes Friedrich, University of Bayreuth (Germany)
}
