% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_ISO_TUN.R
\name{run_MC_ISO_TUN}
\alias{run_MC_ISO_TUN}
\title{Run Monte-Carlo Simulation for Isothermal Measurements for Tunneling Transition}
\usage{
run_MC_ISO_TUN(E, s, T = 200, rho, times, clusters = 10, r = NULL,
  N_e = 200, method = "par", output = "signal", ...)
}
\arguments{
\item{E}{\link{numeric} (\strong{required}): Thermal activation energy of the trap (eV).}

\item{s}{\link{numeric} (\strong{required}): Escape frequency of the trap (s^-1).}

\item{T}{\link{numeric} (\strong{required}): Temperature (deg. C).}

\item{rho}{\link{numeric} (\strong{required}): The calculated dimesionless charge density (normally written Rho') (also defined as (4 *  pi rho/3)^1/3*r) where \code{rho} := the density of recombination centers within the material given in m^3 and \code{r} is _).}

\item{times}{\link{numeric} (\emph{with default}): The specified time within the simulation with the same syntax as the function seq().}

\item{clusters}{\link{numeric} (\emph{with default}): The number of MC runs.}

\item{r}{\link{numeric} (\emph{with default}): The radius of tunneling (dimensionless) .}

\item{N_e}{\link{numeric} (\emph{with default}): The number of electrons.}

\item{method}{\link{character} (\emph{with default}): sequential \code{'seq'} or parallel processing \code{'par'}}

\item{output}{\link{character} (\emph{with default}): output is either the \code{'signal'} (the default) or \code{'remaining_e'} (the remaining
charges, electrons, in the trap)}

\item{\dots}{further arguments}
}
\value{
This function returns a list.
}
\description{
Runs a Monte-Carlo (MC) simulation of isothermally stimulated luminesence (ISO-TL or ITL) using the tunneling (TUN) model. Tunneling refers to the movement of electrons from a trap directly to the recombination center.
}
\details{
\deqn{
I_{TUN}(t) = -dn/dt = A * (n^2 / (r + n))
}

Where in the function \code{n} := \code{n_filled} := \code{N} := \code{N_e} := \code{rho} := \code{rho'} := \code{r_c} := \code{rho'_c}
}
\section{Function version}{
 0.1.0
}

\examples{
\dontrun{
##============================================================================##
## Example 1: Simulate isothermal measurement
##============================================================================##
run_MC_ISO_TUN(
 E = 1.2,
 s = 1e10,
 T = 200,
 rho = 0.007,
 times = 0:5000) \%>\%
  plot_RLumCarlo(legend = TRUE)
}
} 

\section{How to cite}{
Friedrich, J., Kreutzer, S., 2019. run_MC_ISO_TUN(): Run Monte-Carlo Simulation for Isothermal Measurements for Tunneling Transition. Function version 0.1.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-79. 
}

\references{
Pagonis, V. and Kulp, C., 2017. Monte Carlo simulations of tunneling phenomena and nearest neighbor hopping mechanism in feldspars. Journal of Luminescence 181, 114–120. \doi{10.1016/j.jlumin.2016.09.014}

Pagonis, V., Friedrich, J., Discher, M., Müller-Kirschbaum, A., Schlosser, V., Kreutzer, S., Chen, R. and Schmidt, C., 2019. Excited state luminescence signals from a random distribution of defects: A new Monte Carlo simulation approach for feldspar. Journal of Luminescence 207, 266–272. \doi{10.1016/j.jlumin.2018.11.024}

For a discussion of tunneling see:
Aitken, M.J., 1985. Thermoluminescence dating. 276-280. \doi{10.1002/gea.3340020110}
}
\author{
Johannes Friedrich, University of Bayreuth (Germany), Sebastian Kreutzer, IRAMAT-CRP2A,
UMR 5060, CNRS - Univerité Bordeaux Montaigne (France)
}
