% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_TL.R
\name{run_MC_TL}
\alias{run_MC_TL}
\title{Run Monte-Carlo simulation for TL}
\usage{
run_MC_TL(s, E, rho, r_c, times, clusters = 10, N_e = 200,
  delta.r = 0.1, method = "par", output = "signal", ...)
}
\arguments{
\item{s}{\code{\link{list}}}

\item{E}{\code{\link{numeric}}}

\item{rho}{\code{\link{numeric}}}

\item{r_c}{\code{\link{numeric}} (with default)}

\item{times}{\code{\link{vector}} (with default)}

\item{clusters}{\code{\link{numeric}} (with default):}

\item{N_e}{\code{\link{numeric}} (with default):}

\item{delta.r}{\code{\link{numeric}} (with default):}

\item{method}{\code{\link{character}} (with default):}

\item{output}{\code{\link{character}} (with default):}

\item{\dots}{further arguments}
}
\value{
This function returns an \code{\link{array}} with dimension length(times) x length(r) x clusters
}
\description{
Run Monte-Carlo simulation for TL
}
\section{Function version}{
 0.1.0
}

\examples{
\dontrun{
##============================================================================##
## Example 1: Simulate TL measurement
##============================================================================##

times <- seq(200, 500) # time = temperature

run_MC_TL(s = 3.5e12,
          E = 1.45,
          rho = 0.015,
          r_c = 0.85,
          times = times) \%>\%
    plot_RLumCarlo(legend = T)
}
} 

\section{How to cite}{
Friedrich, J., 2019. run_MC_TL(): Run Monte-Carlo simulation for TL. Function version 0.1.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-41. https://CRAN.R-project.org/package=RLumModel
}

\references{
Pagonis 2017
}
\author{
Johannes Friedrich, University of Bayreuth (Germany)
}
