% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_TL_DELOC.R
\name{run_MC_TL_DELOC}
\alias{run_MC_TL_DELOC}
\title{Run Monte-Carlo simulation for TL for delocalized transition}
\usage{
run_MC_TL_DELOC(s, E, times, clusters = 10, N_e = 200,
  n_filled = N_e, R, method = "par", output = "signal", ...)
}
\arguments{
\item{s}{\link{numeric} (\strong{required}): Escape frequency of the trap (s^-1).}

\item{E}{\link{numeric} (\strong{required}): Thermal activation energy of the trap (eV).}

\item{times}{\link{numeric} (\emph{with default}): the number of MC runs.}

\item{clusters}{\link{numeric} (\emph{with default}): the number of clusters.}

\item{N_e}{\link{integer} (\emph{with default}): The number of electrons.}

\item{n_filled}{\link{integer} (\emph{with default}): The number of electron traps that are filled at the beginning of the simulation.}

\item{R}{\link{numeric} (\emph{with default}): The retrapping ratio.}

\item{method}{\link{character} (\emph{with default}): sequential \code{'seq'} or parallel processing \code{'par'}}

\item{output}{\link{character} (\emph{with default}): output is either the \code{'signal'} (the default) or \code{'remaining_e'} (the remaining
charges, electrons, in the trap)}

\item{\dots}{further arguments}
}
\value{
This function returns an \link{array} with dimension length(times) x length(r) x clusters
}
\description{
Runs a Monte-Carlo (MC) simulation of thermo-luminesence (TL) using the one trap one recombination center (OTOR) model. Delocalized refers to involvement of the conduction band.
}
\details{
\deqn{
I_{DELOC}(t) = -dn/dt = p(t) * (n^2 / (NR + n(1-R)))
}

where in the function \code{N} := \code{N_e}:= \code{n} :=\code{n_filled}
}
\section{Function version}{
 0.0.1
}

\examples{
##============================================================================##
## Example 1: Simulate TL
##============================================================================##
\dontrun{
run_MC_TL_DELOC(
 s = 3.5e12,
 E = 1.45,
 R = 1,
 times = 100:450) \%>\%
   plot_RLumCarlo(legend = T)

}

#' @examples
##============================================================================##
## Example 2: Plot multiple TL stimulation TL curves in R with varying params
##============================================================================##
\dontrun{
# define your parameters
times=seq(100,450,1)
s=rep(3.5e12,4)
E=rep(1.45,4)
R<-c(0.7e-6,1e-6,0.01,0.1)
clusters=1000
N_e =c(400, 500, 700, 400)
n_filled =c(400, 500, 300, 70)
method="par"
output ="signal"
col=c(1,2,3,4) # different colours for the individual curves
plot_uncertainty <- c(TRUE,TRUE,TRUE,TRUE)  # do you want to see the uncertainty?
add_TF <- c(FALSE,rep(TRUE, (length(R)-1)))
for (u in 1:length(R)){
 results <-run_MC_TL_DELOC(times=times, s=s[u],E=E[u], clusters =clusters, N_e = N_e[u],
                           n_filled = n_filled[u], R=R[u], method = method, output = output)
 plot_RLumCarlo(results,add=add_TF[u],legend = FALSE, col=col[u], main=" your plot", ylim=c(0,20))
}
legend("topright",ncol=5,cex=0.55,title = "parameters" ,legend=c(paste0("E = ", E),
                                                                paste0("s = ", s),
                                                                paste0("n_filled = ", n_filled),
                                                                paste0("N_e = ", N_e),
                                                                paste0("R = ", R)),  text.col=col)
}

} 

\section{How to cite}{
Kreutzer, S., 2019. run_MC_TL_DELOC(): Run Monte-Carlo simulation for TL for delocalized transition. Function version 0.0.1. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-74. 
}

\references{
Pagonis, V., Friedrich, J., Discher, M., Müller-Kirschbaum, A., Schlosser, V., Kreutzer, S., Chen, R. and Schmidt, C., 2019. Excited state luminescence signals from a random distribution of defects: A new Monte Carlo simulation approach for feldspar. Journal of Luminescence 207, 266–272. \doi{10.1016/j.jlumin.2018.11.024}

Reuven, C. and S. Mckeever, 1997. Theory of thermoluminescence and related phenomena.
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
}
