% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_TL_TUN.R
\name{run_MC_TL_TUN}
\alias{run_MC_TL_TUN}
\title{Run Monte-Carlo simulation for TL for tunneling transition}
\usage{
run_MC_TL_TUN(s, E, rho, r_c, times, clusters = 10, N_e = 200,
  delta.r = 0.1, method = "par", output = "signal", ...)
}
\arguments{
\item{s}{\link{list} (\strong{required}): Escape frequency of the trap (s^-1).}

\item{E}{\link{numeric} (\strong{required}): Thermal activation energy of the trap (eV).}

\item{rho}{\link{numeric} (\strong{required}): The calculated dimesionless Charge density.}

\item{r_c}{\link{numeric} (\emph{with default}): The dimensionless minimal critical radius.}

\item{times}{\link{vector} (\emph{with default}): The number of MC runs.}

\item{clusters}{\link{numeric} (\emph{with default}): The number of clusters.}

\item{N_e}{\link{numeric} (\emph{with default}): The number of electrons}

\item{delta.r}{\link{numeric} (\emph{with default}): The approriate distance interval along the r axis (dimensionless).}

\item{method}{\link{character} (\emph{with default}): ##TODO}

\item{output}{\link{character} (\emph{with default}): ##TODO}

\item{\dots}{further arguments}
}
\value{
This function returns an \code{\link{array}} with dimension length(times) x length(r) x clusters
}
\description{
Runs a Monte-Carlo (MC) simulation of thermo-luminesence (LM-OSL) using the tunneling (TUN) model. Tunneling transitions refers to the direct movement of electrons from a trap directly to the recombination center (\emph{without passing go or collecting $200}).
}
\details{
\deqn{
I_{TUN}(t) = -dn/dt = A * (n^2 / (r + n))
}

where in the function \code{N} := \code{N_e} := \code{rho} := \code{rho'} := \code{r_c} := \code{rho'_c}
}
\section{Function version}{
 0.1.0
}

\examples{
\dontrun{
##============================================================================##
## Example 1: Simulate TL measurement
##============================================================================##

times <- seq(200, 500) # time = temperature

run_MC_TL_TUN(s = 3.5e12,
          E = 1.45,
          rho = 0.015,
          r_c = 0.85,
          times = times) \%>\%
    plot_RLumCarlo(legend = T)
}
} 

\section{How to cite}{
Friedrich, J., 2019. run_MC_TL_TUN(): Run Monte-Carlo simulation for TL for tunneling transition. Function version 0.1.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-67. 
}

\references{
Pagonis, V. and Kulp, C., 2017. Monte Carlo simulations of tunneling phenomena and nearest neighbor hopping mechanism in feldspars. Journal of Luminescence 181, 114–120. \doi{10.1016/j.jlumin.2016.09.014}

Pagonis, V., Friedrich, J., Discher, M., Müller-Kirschbaum, A., Schlosser, V., Kreutzer, S., Chen, R. and Schmidt, C., 2019. Excited state luminescence signals from a random distribution of defects: A new Monte Carlo simulation approach for feldspar. Journal of Luminescence 207, 266–272. \doi{10.1016/j.jlumin.2018.11.024}

for a discussion of tunneling see:
Aitken, M.J., 1985. Thermoluminescence dating. 276-280. \doi{10.1002/gea.3340020110}
}
\author{
Johannes Friedrich, University of Bayreuth (Germany)
}
