% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.RLumCarlo_Model_Output}
\alias{summary.RLumCarlo_Model_Output}
\title{Summarize RLumCarlo Modelling Results}
\usage{
\method{summary}{RLumCarlo_Model_Output}(object, ...)
}
\arguments{
\item{object}{\link{list} of class RLumCarlo_Model_Output: RLumCarlo simulation output object
produced by all}

\item{...}{further arguments passed to the method}
}
\value{
This function returns a \link{data.frame}
}
\description{
Summarize RLumCarlo Modelling results, so that they can be plotted easily
}
\section{Function version}{
 0.1.0
}

\author{
Johannes Friedrich, University of Bayreuth (Germany), Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS-Université Bordeaux Montagine, France
} 

\section{How to cite}{
Friedrich, J., Kreutzer, S., 2019. summary.RLumCarlo_Model_Output(): Summarize RLumCarlo Modelling Results. Function version 0.1.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-44. https://CRAN.R-project.org/package=RLumModel
}

