% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_RLumCarlo.R
\name{calc_RLumCarlo}
\alias{calc_RLumCarlo}
\title{Plot results from Monte-Carlo simulations with RLumCarlo}
\usage{
calc_RLumCarlo(results)
}
\arguments{
\item{results}{\code{\link{array}}:}
}
\value{
This function returns a \code{\link{data.frame}}
}
\description{
Plot results from Monte-Carlo simulations with RLumCarlo
}
\section{Function version}{
 0.0.1 [2017-01-27]
}

\author{
Johannes Friedrich, University of Bayreuth (Germany)
} 

\section{How to cite}{
Friedrich, J., 2019. calc_RLumCarlo(): Plot results from Monte-Carlo simulations with RLumCarlo. Function version 0.0.1 [2017-01-27]. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-16. https://CRAN.R-project.org/package=RLumModel
}

