% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{methods_RLumCarlo}
\alias{methods_RLumCarlo}
\alias{summary.RLumCarlo_Model_Output}
\alias{c.RLumCarlo_Model_Output}
\title{methods_RLumCarlo}
\usage{
\method{summary}{RLumCarlo_Model_Output}(object, verbose = TRUE, ...)

\method{c}{RLumCarlo_Model_Output}(...)
}
\arguments{
\item{object}{\link{list} of class RLumCarlo_Model_Output: RLumCarlo simulation output object
produced by all \code{run_} functions}

\item{verbose}{\link{logical} (\emph{with default}): enable/disable verbose mode}

\item{...}{input objects or further function arguments}
}
\value{
\code{summary()}: This function returns a \link{data.frame}

\code{c()}: function returns a the same as the input objects
}
\description{
Methods for S3-generics implemented for the package 'RLumCarlo'.
}
\examples{

##create object
object <- run_MC_TL_DELOC(
 s = 3.5e12,
 E = 1.45,
 R = 0.1,
 method = 'seq',
 clusters = 100,
 times = 150:350)

##summary
summary(object)

##combine
c(objects,objects)

}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom),
Johannes Friedrich, University of Bayreuth (Germany),
} 

\section{How to cite}{
Kreutzer, S., Friedrich, J., 2020. methods_RLumCarlo(): methods_RLumCarlo. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2020. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence Phenomena. R package version 0.1.7. https://CRAN.R-project.org/package=RLumCarlo
}

\keyword{internal}
