% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{methods_RLumCarlo}
\alias{methods_RLumCarlo}
\alias{summary.RLumCarlo_Model_Output}
\title{methods_RLumCarlo}
\usage{
\method{summary}{RLumCarlo_Model_Output}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{\link{list} of class RLumCarlo_Model_Output: RLumCarlo simulation output object
produced by all \code{run_} functions}

\item{verbose}{\link{logical} (\emph{with default}): enable/disable verbose mode}

\item{...}{further arguments passed to the method}
}
\value{
This function returns a \link{data.frame}
}
\description{
Summarize RLumCarlo modelling results for easy plotting
}
\section{Function version}{
 0.1.0
}

\author{
Johannes Friedrich, University of Bayreuth (Germany),
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS-Université Bordeaux Montaigne (France)
}
\keyword{internal}
