% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RLumCarlo.R
\name{plot_RLumCarlo}
\alias{plot_RLumCarlo}
\title{Plot results from Monte-Carlo simulations with RLumCarlo}
\usage{
plot_RLumCarlo(results, times = NULL, norm = FALSE, legend = FALSE,
  add = FALSE, ...)
}
\arguments{
\item{results}{\code{\link{data.frame}}}

\item{times}{\code{\link{vector}} (with default):}

\item{norm}{\code{\link{character}} (with default):}

\item{legend}{\code{\link{logical}} (with default):}

\item{add}{\code{\link{logical}} (with default):}

\item{\dots}{further arguments}
}
\value{
This function returns a graphical output
}
\description{
Plot results from Monte-Carlo simulations with RLumCarlo
}
\section{Function version}{
 0.0.1 [2017-01-27]
}

\author{
Johannes Friedrich, University of Bayreuth (Germany)
}
