% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RLumCarlo.R
\name{plot_RLumCarlo}
\alias{plot_RLumCarlo}
\title{Plot RLumCarlo Monte-Carlo Simulation Results}
\usage{
plot_RLumCarlo(
  object,
  plot_uncertainty = "range",
  norm = FALSE,
  add = FALSE,
  ...
)
}
\arguments{
\item{object}{\link{list} of class \code{RLumCarlo_Model_Output} (\strong{required}): input object to be plotted,
usually the required input object is generated by one of the functions starting with \code{run_}.
Alternatively a list of such objects can be provided.}

\item{plot_uncertainty}{\link{logical} (\emph{with default}): type of the displayed uncertainty. Allowed
values are \code{range}, \code{sd} (standard deviation) and \code{var} (variance). \code{NULL} disables the uncertainty
visualisation.}

\item{norm}{\link{logical} (\emph{with default}): normalise curve to the highest intensity value}

\item{add}{\link{logical} (\emph{with default}): allows overplotting of results by adding curves to
an existing plot. This argument is handled automatically if \code{object} is of type \link{list}}

\item{\dots}{further arguments that can be passed to control the plot output largely
following the argument names in \link[graphics:plot.default]{graphics::plot.default}. Currently supported
are: \code{xlab}, \code{ylab}, \code{xlim}, \code{ylim}, \code{main}, \code{lwd}, \code{type}, \code{pch}, \code{lty},\code{col}, \code{grid}, \code{legend}.
The arguments \code{lwd}, \code{type}, \code{pch}, \code{lty}, \code{col} can be provided as a vector if \code{object} is a \link{list}}
}
\value{
This function returns a graphical output
}
\description{
Visualise 'RLumCarlo' modelling results without
extracting the values manually. Typically visualised values are the averaged signal
or the number of remaining electrons, with a polygon
indicating modelling uncertainties.
}
\details{
For colouring the curves, the package \link[khroma:khroma-package]{khroma::khroma-package} is used to provide colours
that can be best distinguished, in particular by colour-blind users.
}
\section{Function version}{
 0.1.0
}

\author{
Johannes Friedrich, University of Bayreuth (Germany), Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060,
Université Bordeaux Montaigne (France)
}
\keyword{hplot}
