% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLumCarlo-package.R
\docType{package}
\name{RLumCarlo-package}
\alias{RLumCarlo-package}
\title{Modelling luminescence signals in feldspar}
\description{
\if{html}{
\figure{Logo_RLumCarlo.png}{options: width="50px" alt="https://github.com/R-Lum/RLumCarlo"}\cr
}
}
\details{
\tabular{ll}{ Package: \tab RLumCarlo\cr Type: \tab Package\cr Version:
\tab 0.0.2\cr Date: \tab 2018-08-28 \cr License: \tab GPL-3\cr }
}
\author{
Johannes Friedrich (University of Bayreuth, Germany), Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS-Université Bordeaux Montaigne (France)
}
