% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_CW_IRSL_TUN.R
\encoding{UTF-8}
\name{run_MC_CW_IRSL_TUN}
\alias{run_MC_CW_IRSL_TUN}
\title{Run Monte-Carlo Simulation for CW-IRSL (tunnelling transitions)}
\usage{
run_MC_CW_IRSL_TUN(
  A,
  rho,
  times,
  clusters = 10,
  r_c = 0,
  delta.r = 0.1,
  N_e = 200,
  method = "seq",
  output = "signal",
  ...
)
}
\arguments{
\item{A}{\link{numeric} (\strong{required}): The effective optical excitation rate for the tunnelling process
(\code{s^-1}).}

\item{rho}{\link{numeric} (\strong{required}): The density of recombination centres
(defined as \eqn{\rho}' in Huntley 2006) (dimensionless).}

\item{times}{\link{numeric} (\strong{required}): The sequence of time steps within the simulation (s).}

\item{clusters}{\link{numeric} (\emph{with default}): The number of created clusters for the MC runs. The input can be the output of \link{create_ClusterSystem}. In that case \code{n_filled} indicate absolute numbers of a system.}

\item{r_c}{\link{numeric} (\emph{with default}): Critical distance (>0) that must be provided if the
sample has been thermally and/or optically pretreated. This parameter expresses the fact
that electron-hole pairs within a critical radius \code{r_c} have already recombined.}

\item{delta.r}{\link{numeric} (\emph{with default}): Increments of the dimensionless distance parameter r'}

\item{N_e}{\link{numeric} (\emph{width default}): The total number of electron traps available (dimensionless).
Can be a vector of \code{length(clusters)}, shorter values are recycled.}

\item{method}{\link{character} (\emph{with default}): Sequential \code{'seq'} or parallel \code{'par'}processing. In
the parallel mode the function tries to run the simulation on multiple CPU cores (if available) with
a positive effect on the computation time.}

\item{output}{\link{character} (\emph{with default}): output is either the \code{'signal'} (the default) or
\code{'remaining_e'} (the remaining charges/electrons in the trap)}

\item{\dots}{further arguments, such as \code{cores} to control the number of used CPU cores or \code{verbose} to silence the terminal}
}
\value{
This function returns an object of class \code{RLumCarlo_Model_Output} which
is a \link{list} consisting of an \link{array} with dimension length(times) x length(r) x clusters
and a \link{numeric} time vector.
}
\description{
Runs a Monte-Carlo (MC) simulation of continuous wave infrared stimulated luminescence
(CW-IRSL) using the model for tunnelling transitions. Tunnelling refers to quantum mechanical
tunnelling processes from the excited state of the trap,
into a recombination centre.
}
\details{
\strong{The model}

\deqn{
I_{TUN}(r',t) = -dn/dt = A * exp(-(\rho')^{-1/3} * r')* n (r',t)
}

Where in the function: \cr
A := effective optical excitation rate for the tunnelling process (s^-1) \cr
r' := the dimensionless tunnelling radius \cr
\eqn{\rho}' := \verb{rho'} the dimensionless density of recombination centres (see Huntley (2006)) \cr
t := time (s) \cr
n := the instantaneous number of electrons corresponding to the radius r' at time t
}
\section{Function version}{
 0.2.0
}

\examples{
run_MC_CW_IRSL_TUN(
 A = 0.8,
 rho = 1e-4,
 times = 0:50,
 r_c = 0.05,
 delta.r = 0.1,
 method = "seq",
 clusters = 10,
  output = "signal") \%>\%
 plot_RLumCarlo(norm = TRUE, legend = TRUE)

} 

\section{How to cite}{
Friedrich, J., Kreutzer, S., 2020. run_MC_CW_IRSL_TUN(): Run Monte-Carlo Simulation for CW-IRSL (tunnelling transitions). Function version 0.2.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2020. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence Phenomena. R package version 0.1.7. https://CRAN.R-project.org/package=RLumCarlo
}

\references{
Huntley, D.J., 2006. An explanation of the power-law decay of luminescence.
Journal of Physics: Condensed Matter, 18(4), 1359.

Pagonis, V., Friedrich, J., Discher, M., Müller-Kirschbaum, A., Schlosser, V., Kreutzer, S.,
Chen, R. and Schmidt, C., 2019. Excited state luminescence signals from a random distribution of defects:
A new Monte Carlo simulation approach for feldspar.
Journal of Luminescence 207, 266–272. \doi{10.1016/j.jlumin.2018.11.024}

\strong{Further reading}

Aitken, M.J., 1985. Thermoluminescence dating. Academic Press.

Jain, M., Guralnik, B., Andersen, M.T., 2012. Stimulated luminescence emission from
localized recombination in randomly distributed defects.
Journal of Physics: Condensed Matter 24, 385402.

Chen, R., McKeever, S.W.S., 1997. Theory of Thermoluminescence and Related Phenomena.
WORLD SCIENTIFIC. \doi{10.1142/2781}
}
\author{
Johannes Friedrich, University of Bayreuth (Germany),
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
}
\keyword{data}
\keyword{models}
