% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_CW_IRSL_TUN.R
\name{run_MC_CW_IRSL_TUN}
\alias{run_MC_CW_IRSL_TUN}
\title{Run Monte-Carlo simulation for CW-IRSL}
\usage{
run_MC_CW_IRSL_TUN(A, rho, times, clusters = 10, r = NULL, N_e = 200,
  method = "seq", output = "signal", ...)
}
\arguments{
\item{A}{\link{numeric} (\strong{required}): The transition probability (cm^3/s).}

\item{rho}{\link{numeric} (\strong{required}): The calculated dimesionless Charge density (normally written Rho').}

\item{times}{\link{numeric} (\emph{with default}): The number of MC runs.}

\item{clusters}{\link{numeric} (\emph{with default}): The number of clusters.}

\item{r}{\link{numeric} (\emph{with default}): The retrapping ratio.}

\item{N_e}{\link{numeric} (\emph{with default}): The number of electrons}

\item{method}{\link{character} (\emph{with default}): sequential \code{'seq'} or parallel processing \code{'par'}}

\item{output}{\link{character} (\emph{with default}): output is either the \code{'signal'} (the default) or \code{'remaining_e'} (the remaining
charges, electrons, in the trap)}

\item{\dots}{further arguments}
}
\value{
This function returns a list.
}
\description{
Runs a Monte-Carlo (MC) simulation of constant wave infrared stimulated luminesence (CW-IRSL) using the model.
}
\details{
####equation here please####
}
\section{Function version}{
 0.2.0
}

\examples{
\dontrun{

##============================================================================##
## Example 1: Simulate CW-IRSL measurement
##============================================================================##

run_MC_CW_IRS_TUNL(A = 0.12, rho = 0.003, times = 0:1000) \%>\%
    plot_RLumCarlo(norm = T, legend = T)
}
} 

\section{How to cite}{
Friedrich, J., Kreutzer, S., 2019. run_MC_CW_IRSL_TUN(): Run Monte-Carlo simulation for CW-IRSL. Function version 0.2.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-74. 
}

\references{
Pagonis, V. and Kulp, C. (2017) ‘Monte Carlo simulations of tunneling phenomena and nearest neighbor hopping mechanism in feldspars’, Journal of Luminescence. Elsevier, 181, pp. 114–120. doi: 10.1016/j.jlumin.2016.09.014.
}
\author{
Johannes Friedrich, University of Bayreuth (Germany), Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, Université Bordeaux Montaigne (France)
}
