% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_CW_IRSL_TUN.R
\name{run_MC_CW_IRSL_TUN}
\alias{run_MC_CW_IRSL_TUN}
\title{Run Monte-Carlo simulation for CW-IRSL}
\usage{
run_MC_CW_IRSL_TUN(A, rho, times, clusters = 10, r = NULL, N_e = 200,
  method = "seq", output = "signal", ...)
}
\arguments{
\item{A}{\code{\link{numeric}}}

\item{rho}{\code{\link{numeric}}}

\item{times}{\code{\link{vector}} (with default)}

\item{clusters}{\code{\link{numeric}} (with default):}

\item{r}{\code{\link{numeric}} (with default)}

\item{N_e}{\code{\link{numeric}} (with default):}

\item{method}{\code{\link{character}} (with default):}

\item{output}{\code{\link{character}} (with default):}

\item{\dots}{further arguments}
}
\value{
This function returns a list.
}
\description{
Run Monte-Carlo simulation for CW-IRSL
}
\section{Function version}{
 0.2.0
}

\examples{
\dontrun{

##============================================================================##
## Example 1: Simulate CW-IRSL measurement
##============================================================================##

run_MC_CW_IRS_TUNL(A = 0.12, rho = 0.003, times = 0:1000) \%>\%
    plot_RLumCarlo(norm = T, legend = T)
}
} 

\section{How to cite}{
Friedrich, J., Kreutzer, S., 2019. run_MC_CW_IRSL_TUN(): Run Monte-Carlo simulation for CW-IRSL. Function version 0.2.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-54. https://CRAN.R-project.org/package=RLumModel
}

\references{
Pagonis 2017
}
\author{
Johannes Friedrich, University of Bayreuth (Germany), Sebastian Kreutzer, IRAMAT-CRP2A, Université Bordeaux Montaigne (France)
}
