% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_ISO.R
\name{run_MC_ISO}
\alias{run_MC_ISO}
\title{Run Monte-Carlo simulation for isothermal measurements}
\usage{
run_MC_ISO(E, s, T = 200, rho, times, clusters = 10, r = NULL,
  N_e = 200, method = "par", output = "signal", ...)
}
\arguments{
\item{E}{\link{numeric} (\strong{required})}

\item{s}{\link{numeric} (\strong{required})}

\item{T}{\link{numeric} (\strong{required})}

\item{rho}{\link{numeric} (\strong{required})}

\item{times}{\link{numeric} (\emph{with default})}

\item{clusters}{\link{numeric} (\emph{with default}):}

\item{r}{\link{numeric} (\emph{with default})}

\item{N_e}{\link{numeric} (\emph{with default})}

\item{method}{\link{character} (\emph{with default})}

\item{output}{\link{character} (\emph{with default})}

\item{\dots}{further arguments}
}
\value{
This function returns a list.
}
\description{
Run Monte-Carlo simulation for isothermal measurements
}
\section{Function version}{
 0.1.0
}

\examples{
\dontrun{
##============================================================================##
## Example 1: Simulate isothermal measurement
##============================================================================##

times <- seq(0, 5000)
run_MC_ISO(
 E = 1.2,
 s = 1e10,
 T = 200,
 rho = 0.007,
 times = times) \%>\%
  calc_RLumCarlo() \%>\%
  plot_RLumCarlo(legend = T)
}
} 

\section{How to cite}{
Friedrich, J., Kreutzer, S., 2019. run_MC_ISO(): Run Monte-Carlo simulation for isothermal measurements. Function version 0.1.0. In: Friedrich, J., Kreutzer, S., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-10. https://CRAN.R-project.org/package=RLumModel
}

\references{
Pagonis 2017
}
\author{
Johannes Friedrich, University of Bayreuth (Germany), Sebastian Kreutzer, IRAMAT-CRP2A,
UMR 5060, CNRS - Univerité Bordeaux Montaigne (France)
}
