% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_ISO_DELOC.R
\name{run_MC_ISO_DELOC}
\alias{run_MC_ISO_DELOC}
\title{Run Monte-Carlo simulation for ISO-TL for delocalized transition}
\usage{
run_MC_ISO_DELOC(s, E, T = 20, times, clusters = 10, N_e = 200,
  n_filled = N_e, R, method = "par", output = "signal", ...)
}
\arguments{
\item{s}{\link{numeric} (\strong{required}): The frequency factor of the trap (s^-1).}

\item{E}{\link{numeric} (\strong{required}): Thermal activation energy of the trap (eV).}

\item{T}{\link{numeric} (\emph{with default}): Constant stimulation temperature (degrees C).}

\item{times}{\link{numeric} (\emph{with default}): the sequence of temperature steps within the simulation (s).}

\item{clusters}{\link{numeric} (\emph{with default}): The number of MC runs (unitless).}

\item{N_e}{\link{integer} (\emph{with default}): The total number of electron traps available (unitless).}

\item{n_filled}{\link{integer} (\emph{with default}): The number of filled electron traps at the beginning of the simulation (unitless).}

\item{R}{\link{numeric} (\emph{with default}): The delocalized retrapping ratio (unitless).}

\item{method}{\link{character} (\emph{with default}): sequential \code{'seq'} or parallel processing \code{'par'}}

\item{output}{\link{character} (\emph{with default}): output is either the \code{'signal'} (the default) or \code{'remaining_e'} (the remaining
charges, electrons, in the trap)}

\item{\dots}{further arguments}
}
\value{
This function returns an \link{array} with dimension length(times) x length(r) x clusters
}
\description{
Runs a Monte-Carlo (MC) simulation of isothermally stimulated luminesence (ISO-TL or ITL) using the one trap one recombination center (OTOR) model. Delocalized refers to involvement of the conduction band.
}
\details{
\deqn{
I_{DELOC}(t) = -dn/dt =(s * e^-E/k_{b} * T_{ISO}) * (n^2 / (NR + n(1-R)))
}

Where in the function: \cr
\code{t} := \code{Time} \cr
\code{e}:= Exponentional function \cr
\eqn{k_{B}} := Boltzmann constant \cr
\eqn{T_{ISO}} = \code{T} \cr
\code{n} := \code{n_filled} \cr
\code{N} := \code{N_e}
}
\section{Function version}{
 0.0.1
}

\examples{
##============================================================================##
## Example 1: Single Plot for Monte-Carlo (MC) simulations for delocalized ISO-TL
##============================================================================##
\dontrun{
run_MC_ISO_DELOC(
 s = 3.5e12,
 E = 1.45,
 T = 200,
 R = 1,
 times = 0:10000
) \%>\%
 #Plot results of the MC simulation
plot_RLumCarlo(legend = T)

}


} 

\section{How to cite}{
Kreutzer, S., 2019. run_MC_ISO_DELOC(): Run Monte-Carlo simulation for ISO-TL for delocalized transition. Function version 0.0.1. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-106. 
}

\references{
Pagonis, V., Friedrich, J., Discher, M., Müller-Kirschbaum, A., Schlosser, V., Kreutzer, S., Chen, R. and Schmidt, C., 2019. Excited state luminescence signals from a random distribution of defects: A new Monte Carlo simulation approach for feldspar. Journal of Luminescence 207, 266–272. \doi{10.1016/j.jlumin.2018.11.024}

Reuven, C. and S. Mckeever, 1997. Theory of thermoluminescence and related phenomena.
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
}
