% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_ISO_LOC.R
\name{run_MC_ISO_LOC}
\alias{run_MC_ISO_LOC}
\title{Run Monte-Carlo simulation for ITL for localised transition}
\usage{
run_MC_ISO_LOC(s, E, T = 20, times, clusters = 10, n_filled = 100, r,
  method = "par", output = "signal", ...)
}
\arguments{
\item{s}{\link{numeric} (\strong{required})}

\item{E}{\link{numeric} (\strong{required})}

\item{T}{\link{numeric} (\emph{with default})}

\item{times}{\link{numeric} (\emph{with default}):}

\item{clusters}{\link{numeric} (\emph{with default}):}

\item{n_filled}{\link{integer} (\emph{with default}):}

\item{r}{\link{numeric} (\emph{with default}):}

\item{method}{\link{character} (\emph{with default}):}

\item{output}{\link{character} (\emph{with default}):}

\item{\dots}{further arguments}
}
\value{
This function returns an \link{array} with dimension length(times) x length(r) x clusters
}
\description{
##TODO
}
\details{
\deqn{
I_{LOC}(t) = -dn/dt = p(t) * (n^2 / (r + n))
}
}
\section{Function version}{
 0.0.1
}

\examples{
##============================================================================##
## Example 1: Simulate ITL
##============================================================================##
\dontrun{
run_MC_ISO_LOC(
 s = 3.5e12,
 E = 1.45,
 T = 200,
 r = 1,
 times = 0:10000) \%>\%
   calc_RLumCarlo() \%>\%
   plot_RLumCarlo(legend = T)

}

} 

\section{How to cite}{
Kreutzer, S., 2019. run_MC_ISO_LOC(): Run Monte-Carlo simulation for ITL for localised transition. Function version 0.0.1. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-16. https://CRAN.R-project.org/package=RLumModel
}

\references{
##TODO
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
}
