% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_ISO_TUN.R
\name{run_MC_ISO_TUN}
\alias{run_MC_ISO_TUN}
\title{Run Monte-Carlo simulation for isothermal measurements}
\usage{
run_MC_ISO_TUN(E, s, T = 200, rho, times, clusters = 10, r = NULL,
  N_e = 200, method = "par", output = "signal", ...)
}
\arguments{
\item{E}{\link{numeric} (\strong{required}): Thermal activation energy of the trap (eV).}

\item{s}{\link{numeric} (\strong{required}): Escape frequency of the trap (s^-1).}

\item{T}{\link{numeric} (\strong{required}): Temperature (deg. C).}

\item{rho}{\link{numeric} (\strong{required}): The calculated dimesionless Charge density (normally written Rho').}

\item{times}{\link{numeric} (\emph{with default}): The number of MC runs.}

\item{clusters}{\link{numeric} (\emph{with default}): The number of clusters.}

\item{r}{\link{numeric} (\emph{with default}): The retrapping ratio.}

\item{N_e}{\link{numeric} (\emph{with default}): The number of electrons.}

\item{method}{\link{character} (\emph{with default})}

\item{output}{\link{character} (\emph{with default})}

\item{\dots}{further arguments}
}
\value{
This function returns a list.
}
\description{
Runs a Monte-Carlo (MC) simulation of isothermally stimulated luminesence (ISO-TL or ITL) using the .
}
\details{
\deqn{
I_{LOC}(t) = -dn/dt = A * (n^2 / (r + n))
}
}
\section{Function version}{
 0.1.0
}

\examples{
\dontrun{
##============================================================================##
## Example 1: Simulate isothermal measurement
##============================================================================##

times <- seq(0, 5000)
run_MC_ISO_TUN(
 E = 1.2,
 s = 1e10,
 T = 200,
 rho = 0.007,
 times = times) \%>\%
  plot_RLumCarlo(legend = T)
}
} 

\section{How to cite}{
Friedrich, J., Kreutzer, S., 2019. run_MC_ISO_TUN(): Run Monte-Carlo simulation for isothermal measurements. Function version 0.1.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-61. 
}

\references{
Pagonis, V. and Kulp, C., 2017. ‘Monte Carlo simulations of tunneling phenomena and nearest neighbor hopping mechanism in feldspars’, Journal of Luminescence 181, 114–120. \doi{10.1016/j.jlumin.2016.09.014}
}
\author{
Johannes Friedrich, University of Bayreuth (Germany), Sebastian Kreutzer, IRAMAT-CRP2A,
UMR 5060, CNRS - Univerité Bordeaux Montaigne (France)
}
