% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_ISO_TUN.R
\name{run_MC_ISO_TUN}
\alias{run_MC_ISO_TUN}
\title{Run Monte-Carlo Simulation for ISO-TL for tunneling transition}
\usage{
run_MC_ISO_TUN(E, s, T = 200, rho, times, clusters = 10, r_c = 0,
  delta.r = 0.1, N_e = 200L, method = "par", output = "signal",
  ...)
}
\arguments{
\item{E}{\link{numeric} (\strong{required}): Thermal activation energy of the trap (eV).}

\item{s}{\link{numeric} (\strong{required}): Frequency factor of the trap (s^-1).}

\item{T}{\link{numeric} (\strong{required}): Constant stimulation temperature (degrees C).}

\item{rho}{\link{numeric} (\strong{required}): The density of recombination centers (defined as rho' in Huntley 2006) (unitless).}

\item{times}{\link{numeric} (\emph{with default}): The sequence of temperature steps within the simulation (s).}

\item{clusters}{\link{numeric} (\emph{with default}): The number of MC runs (unitless).}

\item{r_c}{\link{numeric} (\emph{with default}): The radius of tunneling (dimensionless)}

\item{delta.r}{\link{numeric} (\emph{with default}):}

\item{N_e}{\link{numeric} (\emph{width default}): The total number of electron traps available (unitless).}

\item{method}{\link{character} (\emph{with default}): sequential \code{'seq'} or parallel processing \code{'par'}}

\item{output}{\link{character} (\emph{with default}): output is either the \code{'signal'} (the default) or \code{'remaining_e'} (the remaining
charges, electrons, in the trap)}

\item{\dots}{further arguments}

\item{r}{\link{numeric} (\emph{with default}): The radius of tunneling (unitless).}
}
\value{
This function returns a list.
}
\description{
Runs a Monte-Carlo (MC) simulation of isothermally stimulated luminesence (ISO-TL or ITL) using the tunneling (TUN) model. Tunneling refers to the direct movement of electrons from a trap directly to the recombination center.
}
\details{
\deqn{
p(t) = s * e ^ (-E / k_{B} * T) * e^{(-r' / \rho'^{1 / 3})}
}

\deqn{
I_{TUN}(t) = 3 * n * p(t) *  (r')^{2 * e^{(-r'^3)}}
}

Where in the function: \cr
\code{p(t)} := The experimental stimulation mode \cr
\code{e}:= Exponentional function \cr
\eqn{k_{B}} := Boltzmann constant \cr
\code{r} := \code{r} \cr
\eqn{\rho} := \code{rho} \cr
\code{t} := \code{Time} \cr
\code{n} := The Instantaneous number of electrons \cr
\code{n} := \code{n_filled} \cr
\code{t}:= \code{times}
}
\section{Function version}{
 0.1.0
}

\examples{
##============================================================================##
## Example 1: Single Plot for Monte-Carlo (MC) simulations for tunneling ISO-TL
##============================================================================##
\dontrun{
run_MC_ISO_TUN(
 E = 0.002,
 s = 1e16,
 T = 50,
 rho = 1e-4,
 times = 0:10,
 clusters = 100,
 N_e = 2,
 r_c = 1e-4,
 delta.r = 0.5,
 method = "par"
) \%>\%
 ##Plot results of the MC simulation
plot_RLumCarlo(legend = TRUE)

}

} 

\section{How to cite}{
Friedrich, J., Kreutzer, S., 2019. run_MC_ISO_TUN(): Run Monte-Carlo Simulation for ISO-TL for tunneling transition. Function version 0.1.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-102. 
}

\references{
Huntley, D.J., 2006. An explanation of the power-law decay of luminescence. Journal of Physics: Condensed Matter, 18(4), 1359.\doi{10.1088/0953-8984/18/4/020}

Pagonis, V. and Kulp, C., 2017. Monte Carlo simulations of tunneling phenomena and nearest neighbor hopping mechanism in feldspars. Journal of Luminescence 181, 114–120. \doi{10.1016/j.jlumin.2016.09.014}

Pagonis, V., Friedrich, J., Discher, M., Müller-Kirschbaum, A., Schlosser, V., Kreutzer, S., Chen, R. and Schmidt, C., 2019. Excited state luminescence signals from a random distribution of defects: A new Monte Carlo simulation approach for feldspar. Journal of Luminescence 207, 266–272. \doi{10.1016/j.jlumin.2018.11.024}

\strong{Further reading}
Aitken, M.J., 1985. Thermoluminescence dating. 276-280. \doi{10.1002/gea.3340020110}
}
\author{
Johannes Friedrich, University of Bayreuth (Germany), Sebastian Kreutzer, IRAMAT-CRP2A,
UMR 5060, CNRS - Univerité Bordeaux Montaigne (France)
}
