% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_LM_OSL.R
\name{run_MC_LM_OSL}
\alias{run_MC_LM_OSL}
\title{Run Monte-Carlo simulation for LM-OSL}
\usage{
run_MC_LM_OSL(A, rho, times, clusters = 10, r = NULL, delta.r = 0.1,
  N_e = 200, method = "par", output = "signal", ...)
}
\arguments{
\item{A}{\code{\link{numeric}}}

\item{rho}{\code{\link{numeric}}}

\item{times}{\code{\link{vector}} (with default)}

\item{clusters}{\code{\link{numeric}} (with default):}

\item{r}{\code{\link{numeric}} (with default):}

\item{delta.r}{\code{\link{numeric}} (with default):}

\item{N_e}{\code{\link{numeric}} (with default):}

\item{method}{\code{\link{character}} (with default):}

\item{output}{\code{\link{character}} (with default):}

\item{\dots}{further arguments}
}
\value{
This function returns a list.
}
\description{
Run Monte-Carlo simulation for LM-OSL
}
\section{Function version}{
 0.0.1 [2017-01-27]
}

\examples{

\dontrun{

##TODO: Primary example, should be verified
run_MC_LM_OSL(A = 10000, rho = 0.0001, times = 1:100, clusters = 10, r = NULL,
 delta.r = 0.1,
 N_e = 200, method = "par", output = "signal") \%>\%
 calc_RLumCarlo() \%>\%
 plot_RLumCarlo(norm = T)

}

} 

\section{How to cite}{
Friedrich, J., 2019. run_MC_LM_OSL(): Run Monte-Carlo simulation for LM-OSL. Function version 0.0.1 [2017-01-27]. In: Friedrich, J., Kreutzer, S., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-10. https://CRAN.R-project.org/package=RLumModel
}

\references{
Pagonis 2017
}
\author{
Johannes Friedrich, University of Bayreuth (Germany)
}
