% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_LM_OSL_LOC.R
\encoding{UTF-8}
\name{run_MC_LM_OSL_LOC}
\alias{run_MC_LM_OSL_LOC}
\title{Run Monte-Carlo Simulation for LM-OSL (localized transitions)}
\usage{
run_MC_LM_OSL_LOC(
  A,
  times,
  clusters = 10,
  n_filled = 100,
  r,
  method = "par",
  output = "signal",
  ...
)
}
\arguments{
\item{A}{\link{numeric} (\strong{required}): The optical excitation rate from the ground state into the excited
state of the trap (s^-1)}

\item{times}{\link{numeric} (\strong{required}): The sequence of time steps within the simulation (s)}

\item{clusters}{\link{numeric} (\emph{with default}): The number of created clusters for the MC runs. The input can be the output of \link{create_ClusterSystem}. In that case \code{n_filled} indicate absolute numbers of a system.}

\item{n_filled}{\link{integer} (\emph{with default}): The number of filled electron traps at the
beginning of the simulation (dimensionless). Can be a vector of \code{length(clusters)}, shorter values are recycled.}

\item{r}{\link{numeric} (\strong{required}): The retrapping ratio for localized transitions}

\item{method}{\link{character} (\emph{with default}): Sequential \code{'seq'} or parallel \code{'par'}processing. In
the parallel mode the function tries to run the simulation on multiple CPU cores (if available) with
a positive effect on the computation time.}

\item{output}{\link{character} (\emph{with default}): output is either the \code{'signal'} (the default)
or \code{'remaining_e'} (the remaining charges, electrons, in the trap)}

\item{\dots}{further arguments, such as \code{cores} to control the number of used CPU cores or \code{verbose} to silence the terminal}
}
\value{
This function returns an object of class \code{RLumCarlo_Model_Output} which
is a \link{list} consisting of an \link{array} with dimension length(times) x clusters
and a \link{numeric} time vector.
}
\description{
Runs a Monte-Carlo (MC) simulation of linearly modulated optically stimulated
luminescence (LM-OSL) using the generalized one trap (GOT) model. Localized transitions refer to
transitions which do not involve the conduction or valence band. These transitions take place
between the ground state and an excited state of the trap, and also involve
a an energy state of the recombination centre.
}
\details{
\strong{The model}

\deqn{
I_{LOC}(t) = -dn/dt = (A * t/P) * (n^2 / (r + n))
}

Where in the function: \cr
A := optical excitation rate from the ground state into the excited state of the trap (1/s)\cr
P := total excitation time (s) \cr
t := time (s) \cr
n := \code{n_filled}, the instantaneous number of electrons \cr
r := the retrapping ratio for localized transitions
}
\section{Function version}{
 0.1.0
}

\examples{
## short example
run_MC_LM_OSL_LOC(
 A = 1,
 times = 0:40,
 clusters = 10,
 n_filled = 100,
 r = 1e-7,
 method = "seq",
 output = "signal") \%>\%
plot_RLumCarlo(legend = TRUE)

\dontrun{
## the long (meaningful) example
results <- run_MC_LM_OSL_LOC(
 A = 1,
 times = 0:100,
 clusters = 100,
 n_filled = 100,
 r = 1e-7,
 method = "par",
 output = "signal")

## plot
plot_RLumCarlo(results, legend = TRUE)
}

} 

\section{How to cite}{
Kreutzer, S., 2020. run_MC_LM_OSL_LOC(): Run Monte-Carlo Simulation for LM-OSL (localized transitions). Function version 0.1.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2020. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence Phenomena. R package version 0.1.7. https://CRAN.R-project.org/package=RLumCarlo
}

\references{
Pagonis, V., Friedrich, J., Discher, M., Müller-Kirschbaum, A., Schlosser, V.,
Kreutzer, S., Chen, R. and Schmidt, C., 2019. Excited state luminescence signals
from a random distribution of defects: A new Monte Carlo simulation approach for feldspar.
Journal of Luminescence 207, 266–272. \doi{10.1016/j.jlumin.2018.11.024}
}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
}
\keyword{data}
\keyword{models}
