% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_LM_OSL_LOC.R
\name{run_MC_LM_OSL_LOC}
\alias{run_MC_LM_OSL_LOC}
\title{Run Monte-Carlo simulation for LM-OSL for localised transition}
\usage{
run_MC_LM_OSL_LOC(A, times, clusters = 10, n_filled = 100, r,
  method = "par", output = "signal", ...)
}
\arguments{
\item{A}{\link{numeric} (\strong{required})}

\item{times}{\link{numeric} (\emph{with default}):}

\item{clusters}{\link{numeric} (\emph{with default}):}

\item{n_filled}{\link{integer} (\emph{with default}):}

\item{r}{\link{numeric} (\emph{with default}):}

\item{method}{\link{character} (\emph{with default}):}

\item{output}{\link{character} (\emph{with default}):}

\item{\dots}{further arguments}
}
\value{
This function returns an \link{array} with dimension length(times) x length(r) x clusters
}
\description{
##TODO
}
\details{
\deqn{
I_{LOC}(t) = -dn/dt = A * (n^2 / (r + n))
}
}
\section{Function version}{
 0.0.1
}

\examples{
##============================================================================##
## Example 1: Simulate LM-OSL
##============================================================================##
\dontrun{
run_MC_LM_OSL_LOC(
 A = 0.12,
 r = 1,
 times = 0:100) \%>\%
   plot_RLumCarlo(legend = T)

}

} 

\section{How to cite}{
Kreutzer, S., 2019. run_MC_LM_OSL_LOC(): Run Monte-Carlo simulation for LM-OSL for localised transition. Function version 0.0.1. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-47. https://CRAN.R-project.org/package=RLumModel
}

\references{
##TODO
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
}
