% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_LM_OSL_TUN.R
\name{run_MC_LM_OSL_TUN}
\alias{run_MC_LM_OSL_TUN}
\title{Run Monte-Carlo simulation for LM-OSL}
\usage{
run_MC_LM_OSL_TUN(A, rho, times, clusters = 10, r_c = 0,
  delta.r = 0.1, N_e = 200, method = "par", output = "signal", ...)
}
\arguments{
\item{A}{\link{numeric} (\strong{required}): The transition probaility (cm^3/s).}

\item{rho}{\link{numeric} (\strong{required}): The calculated dimesionless Charge density (normally written Rho').}

\item{times}{\link{vector} (\emph{with default}): The number of MC runs.}

\item{clusters}{\link{numeric} (\emph{with default}): The number of clusters.}

\item{r_c}{\link{numeric} (\emph{with default}): The retrapping ratio.}

\item{delta.r}{\link{numeric} (\emph{with default}):}

\item{N_e}{\link{numeric} (\emph{with default}): The number of electrons.}

\item{method}{\link{character} (\emph{with default}): sequential \code{'seq'} or parallel processing \code{'par'}}

\item{output}{\link{character} (\emph{with default}): output is either the \code{'signal'} (the default) or \code{'remaining_e'} (the remaining
charges, electrons, in the trap)}

\item{\dots}{further arguments}
}
\value{
This function returns a list.
}
\description{
Run Monte-Carlo simulation for LM-OSL
}
\details{
ADD EQUATION
}
\section{Function version}{
 0.1.0
}

\examples{

\dontrun{

##TODO: Primary example, should be verified
run_MC_LM_OSL_TUN(A = 10000, rho = 0.0001, times = 1:100, clusters = 10, r = NULL,
 delta.r = 0.1,
 N_e = 200, method = "par", output = "signal") \%>\%
 plot_RLumCarlo(norm = T)

}

} 

\section{How to cite}{
Friedrich, J., 2019. run_MC_LM_OSL_TUN(): Run Monte-Carlo simulation for LM-OSL. Function version 0.1.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-85. 
}

\references{
Pagonis, V. and Kulp, C. (2017) ‘Monte Carlo simulations of tunneling phenomena and nearest neighbor hopping mechanism in feldspars’, Journal of Luminescence. Elsevier, 181, pp. 114–120. doi: 10.1016/j.jlumin.2016.09.014.
}
\author{
Johannes Friedrich, University of Bayreuth (Germany)
}
