% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_TL_DELOC.R
\name{run_MC_TL_DELOC}
\alias{run_MC_TL_DELOC}
\title{Run Monte-Carlo simulation for TL for GOT model}
\usage{
run_MC_TL_DELOC(s, E, times, clusters = 10, N_e = 200,
  n_filled = N_e, R, method = "par", output = "signal", ...)
}
\arguments{
\item{s}{\link{numeric} (\emph{required})}

\item{E}{\link{numeric} (\emph{required})}

\item{times}{\link{numeric} (with default)}

\item{clusters}{\link{numeric} (with default):}

\item{N_e}{\link{integer} (with default)}

\item{n_filled}{\link{integer} (with default)}

\item{R}{\link{numeric} (with default):}

\item{method}{\link{character} (with default):}

\item{output}{\link{character} (with default):}

\item{\dots}{further arguments}
}
\value{
This function returns an \link{array} with dimension length(times) x length(r) x clusters
}
\description{
##TODO
}
\details{
\deqn{
I_{DELOC}(t) = -dn/dt = p(t) * (n^2 / (NR + n(1-R)))
}
}
\section{Function version}{
 0.0.1
}

\examples{
##============================================================================##
## Example 1: Simulate TL
##============================================================================##
\dontrun{
run_MC_TL_DELOC(
 s = 3.5e12,
 E = 1.45,
 R = 1,
 times = 100:450) \%>\%
   calc_RLumCarlo() \%>\%
   plot_RLumCarlo(legend = T)

}

} 

\section{How to cite}{
Kreutzer, S., 2019. run_MC_TL_DELOC(): Run Monte-Carlo simulation for TL for GOT model. Function version 0.0.1. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-15. https://CRAN.R-project.org/package=RLumModel
}

\references{
##TODO
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
}
