% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_TL_LOC.R
\name{run_MC_TL_LOC}
\alias{run_MC_TL_LOC}
\title{Run Monte-Carlo Simulation for TL (localized transitions)}
\usage{
run_MC_TL_LOC(
  s,
  E,
  times,
  clusters = 10,
  n_filled = 100,
  r,
  method = "par",
  output = "signal",
  ...
)
}
\arguments{
\item{s}{\link{numeric} (\strong{required}): The frequency factor of the trap (s^-1)}

\item{E}{\link{numeric} (\strong{required}): Thermal activation energy of the trap (eV)}

\item{times}{\link{numeric} (\emph{with default}): The sequence of temperature steps within the simulation (s)}

\item{clusters}{\link{numeric} (\emph{with default}): The number of MC run (unitless)}

\item{n_filled}{\link{integer} (\emph{with default}): The number of filled electron traps at
the beginning of the simulation (unitless)}

\item{r}{\link{numeric} (\emph{with default}): The localized retrapping ratio (unitless)}

\item{method}{\link{character} (\emph{with default}): Sequential \code{'seq'} or parallel \code{'par'}processing. In
the parallel mode the function tries to run the simulation on multiple CPU cores (if available) with
a positive effect on the computation time.}

\item{output}{\link{character} (\emph{with default}): output is either the \code{'signal'} (the default) or
\code{'remaining_e'} (the remaining charges/electrons in the trap)}

\item{\dots}{further arguments}
}
\value{
This function returns an object of class \code{RLumCarlo_Model_Output} which
is a \link{list} consisting of an \link{array} with dimension length(times) x clusters
and a \link{numeric} time vector.
}
\description{
Runs a Monte-Carlo (MC) simulation of thermoluminesence (TL) using
the generalized one trap (GOT) model. Localized transitions refer to transitions
which do not involve the conduction or valence band. These transitions take place between the
ground state and an excited state of the trapped charge, and also involve an excited
state of the recombination center.
}
\details{
\strong{The model}

\deqn{
I_{LOC}(t) = -dn/dt = (s * exp(-E/(k_{B} * T))) * (n^2 / (r + n))
}

Where in the function: \cr
E := the thermal activation energy (eV) \cr
s := the frequency factor for the trap (s^-1) \cr
t := time (s) \cr
\eqn{k_{B}} := Boltzmann constant \cr
T := temperature (degrees C) \cr
n := the instantaneous number of electrons \cr
r := the retrapping ratio for localized transitions
}
\section{Function version}{
 0.1.0
}

\examples{
## the short example
run_MC_TL_LOC(
 s = 1e14,
 E = 0.9,
 times = 50:100,
 method = "seq",
 clusters = 2,
 r = 1e4) \%>\%
plot_RLumCarlo()

\dontrun{
## the long (meaningful) example
results <- run_MC_TL_LOC(
 s = 1e14,
 E = 0.9,
 times = 50:100,
 method = "par",
 clusters = 100,
 r = 1e4)

## plot
plot_RLumCarlo(results)

}

} 

\section{How to cite}{
Kreutzer, S., 2019. run_MC_TL_LOC(): Run Monte-Carlo Simulation for TL (localized transitions). Function version 0.1.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-131. 
}

\references{
Pagonis, V., Friedrich, J., Discher, M., Müller-Kirschbaum, A., Schlosser, V.,
reutzer, S., Chen, R. and Schmidt, C., 2019. Excited state luminescence signals
from a random distribution of defects: A new Monte Carlo simulation approach for feldspar.
Journal of Luminescence 207, 266–272. \doi{10.1016/j.jlumin.2018.11.024}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
}
\keyword{data}
\keyword{models}
