% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_TL_LOC.R
\name{run_MC_TL_LOC}
\alias{run_MC_TL_LOC}
\title{Run Monte-Carlo simulation for TL for localized transition}
\usage{
run_MC_TL_LOC(s, E, times, clusters = 10, n_filled = 100, r,
  method = "par", output = "signal", ...)
}
\arguments{
\item{s}{\link{numeric} (\strong{required}): The frequency factor of the trap (s^-1).}

\item{E}{\link{numeric} (\strong{required}): Thermal activation energy of the trap (eV).}

\item{times}{\link{numeric} (\emph{with default}): The sequence of temperature steps within the simulation (s).}

\item{clusters}{\link{numeric} (\emph{with default}): The number of MC run (unitless).}

\item{n_filled}{\link{integer} (\emph{with default}): The number of filled electron traps at the beginning of the simulation (unitless).}

\item{r}{\link{numeric} (\emph{with default}): The localized retrapping ratio (unitless).}

\item{method}{\link{character} (\emph{with default}): sequential \code{'seq'} or parallel processing \code{'par'}}

\item{output}{\link{character} (\emph{with default}): output is either the \code{'signal'} (the default) or \code{'remaining_e'} (the remaining
charges, electrons, in the trap)}

\item{\dots}{further arguments}
}
\value{
This function returns an \link{array} with dimension length(times) x length(r) x clusters
}
\description{
Runs a Monte-Carlo (MC) simulation of thermo-luminesence (TL) using the generalized one trap (GOT) model. Localized refers to excitation of an electron before it recombines, but without the involvement of the conduction band.
}
\details{
\deqn{
I_{LOC}(t) = -dn/dt = (s * e^-E/k_b * T) * (n^2 / (r + n))
}

Where in the function: \cr
\code{t} := \code{Time} \cr
\code{e}:= Exponentional function \cr
\eqn{k_{B}} := Boltzmann constant \cr
\code{T} := \code{Temperature} \cr
\code{n} := The Instantaneous number of electrons
}
\section{Function version}{
 0.1.0
}

\examples{
##============================================================================##
## Example 1: Single Plot for Monte-Carlo (MC) simulations for localized TL
##============================================================================## 
\dontrun{
run_MC_TL_LOC(
 s = 1e8,
 E = 0.5,
 times = 0:100,
 clusters = 1e1,
 n_filled = 1,
 r = 1e-7) \%>\%
 #Plot results of the MC simulation
   plot_RLumCarlo(legend = T)
   
}



} 

\section{How to cite}{
Kreutzer, S., 2019. run_MC_TL_LOC(): Run Monte-Carlo simulation for TL for localized transition. Function version 0.1.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-101. 
}

\references{
Pagonis, V., Friedrich, J., Discher, M., Müller-Kirschbaum, A., Schlosser, V., Kreutzer, S., Chen, R. and Schmidt, C., 2019. Excited state luminescence signals from a random distribution of defects: A new Monte Carlo simulation approach for feldspar. Journal of Luminescence 207, 266–272. \doi{10.1016/j.jlumin.2018.11.024}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
}
