% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_TL_LOC.R
\name{run_MC_TL_LOC}
\alias{run_MC_TL_LOC}
\title{Run Monte-Carlo simulation for TL for localised transition}
\usage{
run_MC_TL_LOC(s, E, times, clusters = 10, n_filled = 100, r,
  method = "par", output = "signal", ...)
}
\arguments{
\item{s}{\link{numeric} (\strong{required}): Escape frequency of the trap (s^-1).}

\item{E}{\link{numeric} (\strong{required}): Thermal activation energy of the trap (eV).}

\item{times}{\link{numeric} (\emph{with default}): The number of Mc runs.}

\item{clusters}{\link{numeric} (\emph{with default}): The number of clusters.}

\item{n_filled}{\link{integer} (\emph{with default}): The number of electron traps that are filled at the beginning of the simulation.}

\item{r}{\link{numeric} (\emph{with default}): The retrapping ratio.}

\item{method}{\link{character} (\emph{with default}):}

\item{output}{\link{character} (\emph{with default}):}

\item{\dots}{further arguments}
}
\value{
This function returns an \link{array} with dimension length(times) x length(r) x clusters
}
\description{
Runs a Monte-Carlo (MC) simulation of thermo-luminesence (LM-OSL) using the generalized one trap (GOT) model.
}
\details{
\deqn{
I_{LOC}(t) = -dn/dt = p(t) * (n^2 / (r + n))
}
}
\section{Function version}{
 0.1.0
}

\examples{
##============================================================================##
## Example 1: Simulate TL
##============================================================================##
\dontrun{
run_MC_TL_LOC(
 s = 3.5e12,
 E = 1.45,
 r = 1,
 times = 100:450) \%>\%
   plot_RLumCarlo(legend = T)

}

} 

\section{How to cite}{
Kreutzer, S., 2019. run_MC_TL_LOC(): Run Monte-Carlo simulation for TL for localised transition. Function version 0.1.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-60. https://CRAN.R-project.org/package=RLumModel
}

\references{
Pagonis, V. and Kulp, C. (2017) ‘Monte Carlo simulations of tunneling phenomena and nearest neighbor hopping mechanism in feldspars’, Journal of Luminescence. Elsevier, 181, pp. 114–120. doi: 10.1016/j.jlumin.2016.09.014.
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
}
