% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_TL_TUN.R
\name{run_MC_TL_TUN}
\alias{run_MC_TL_TUN}
\title{Run Monte-Carlo Simulation for TL (tunneling transitions)}
\usage{
run_MC_TL_TUN(
  s,
  E,
  rho,
  r_c = 0,
  times,
  clusters = 10,
  N_e = 200,
  delta.r = 0.1,
  method = "par",
  output = "signal",
  ...
)
}
\arguments{
\item{s}{\link{list} (\strong{required}): The effective frequency factor for the tunneling process (1/s)}

\item{E}{\link{numeric} (\strong{required}): Thermal activation energy of the trap (eV)}

\item{rho}{\link{numeric} (\strong{required}): The dimensionless density of recombination centers
(defined as \eqn{\rho}' in Huntley 2006)}

\item{r_c}{\link{numeric} (\emph{with default}): Critical distance (>0) that is to be used if
the sample has been thermally and/or optically pretreated. This parameter expresses the fact
that electron-hole pairs within a critical radius \code{r_c} have already been recombined.}

\item{times}{\link{vector} (\emph{wih default}): The sequence of time steps within the simulation (s)}

\item{clusters}{\link{numeric} (\emph{with default}): The number of MC runs (unitless)}

\item{N_e}{\link{numeric} (\emph{with default}): The total number of electron traps available (unitless)}

\item{delta.r}{\link{numeric} (\emph{with default}): The increments of \code{r_c} (unitless)}

\item{method}{\link{character} (\emph{with default}): sequential \code{'seq'} or parallel processing \code{'par'}.}

\item{output}{\link{character} (\emph{with default}): output is either the \code{'signal'} (the default)
or \code{'remaining_e'} (the remaining charges, electrons, in the trap)}

\item{\dots}{further arguments}
}
\value{
This function returns an object of class \code{RLumCarlo_Model_Output} which
is a \link{list} consisting of an \link{array} with dimension length(times) x length(r) x clusters
and a \link{numeric} time vector.
}
\description{
Runs a Monte-Carlo (MC) simulation of thermo-luminesence (TL) caused by
tunnelling (TUN) transitions.  Tunneling refers to quantum mechanical
tunneling processes from the excited state of the trapped charge,
into a state of the recombination center.
}
\details{
\strong{The model}

\deqn{
I_{TUN}(r',t) = -dn/dt = (s * exp(-E/(k_{B} * T))) * exp(-(\rho')^(-1/3) * r') * n(r',t)
}

Where in the function: \cr
s := effective frequency for the tunneling process (1/s) \cr
E := thermal activation energy (eV) \cr
\eqn{k_{B}} := Boltzmann constant \cr
T := temperature \cr
r' := \code{r} electron-hole distance (unitless) \cr
\eqn{\rho}' := Density of recombination centers \cr
t := time (s) \cr
n := The instantaneous number of electrons
}
\section{Function version}{
 0.1.0
}

\examples{
## the short example
run_MC_TL_TUN(
 s = 1e12,
 E = 0.9,
 rho = 1,
 r_c = 1,
 times = 80:120,
 clusters = 2,
 method = 'seq',
 delta.r = 1e-1) \%>\%
plot_RLumCarlo()

\dontrun{
## the long (meaningful example)
run_MC_TL_TUN(
 s = 1e12,
 E = 0.9,
 rho = 0.01,
 r_c = 0.1,
 times = 80:220,
 clusters = 100,
 method = 'par',
 delta.r = 1e-1) %>%
plot_RLumCarlo()
}

} 

\section{How to cite}{
Friedrich, J., Kreutzer, S., 2019. run_MC_TL_TUN(): Run Monte-Carlo Simulation for TL (tunneling transitions). Function version 0.1.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-127. 
}

\references{
Huntley, D.J., 2006. An explanation of the power-law decay of luminescence.
Journal of Physics: Condensed Matter, 18(4), 1359.\doi{10.1088/0953-8984/18/4/020}

Pagonis, V. and Kulp, C., 2017. Monte Carlo simulations of tunneling
phenomena and nearest neighbor hopping mechanism in feldspars.
Journal of Luminescence 181, 114–120. \doi{10.1016/j.jlumin.2016.09.014}

Pagonis, V., Friedrich, J., Discher, M., Müller-Kirschbaum, A., Schlosser, V., Kreutzer, S.,
Chen, R. and Schmidt, C., 2019. Excited state luminescence signals from a random
distribution of defects: A new Monte Carlo simulation approach for feldspar.
Journal of Luminescence 207, 266–272. \doi{10.1016/j.jlumin.2018.11.024}

\strong{Further reading}
Aitken, M.J., 1985. Thermoluminescence dating. Academic Press.

Jain, M., Guralnik, B., Andersen, M.T., 2012. Stimulated luminescence emission from
localized recombination in randomly distributed defects.
J. Phys.: Condens. Matter 24, 385402. \doi{10.1088/0953-8984/24/38/385402}
}
\author{
Johannes Friedrich, University of Bayreuth (Germany), Sebastian Kreutzer,
IRAMAT-CRP2A, UMR 5060, Université Bordeaux Montaigne (France)
}
\keyword{data}
\keyword{models}
