% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC_TL_TUN.R
\name{run_MC_TL_TUN}
\alias{run_MC_TL_TUN}
\title{Run Monte-Carlo Simulation for TL for tunnelling transition}
\usage{
run_MC_TL_TUN(s, E, rho, r_c = 0, times, clusters = 10, N_e = 200,
  delta.r = 0.1, method = "par", output = "signal", ...)
}
\arguments{
\item{s}{\link{list} (\strong{required}): The frequency factor of the trap (s^-1).}

\item{E}{\link{numeric} (\strong{required}): Thermal activation energy of the trap (eV).}

\item{rho}{\link{numeric} (\strong{required}): The density of recombination centers (defined as rho' in Huntley 2006) (unitless).}

\item{r_c}{\link{numeric} (\emph{with default}): Distance parameter (radius of tunneling) (unitless).}

\item{times}{\link{vector} (\emph{with default}): The sequence of temperature steps within the simulation (s).}

\item{clusters}{\link{numeric} (\emph{with default}): The number of MC runs (unitless).}

\item{N_e}{\link{numeric} (\emph{with default}): The total number of electron traps available (unitless).}

\item{delta.r}{\link{numeric} (\emph{with default}): The  increments of r_c (unitless).}

\item{method}{\link{character} (\emph{with default}): sequential \code{'seq'} or parallel processing \code{'par'}}

\item{output}{\link{character} (\emph{with default}): output is either the \code{'signal'} (the default) or \code{'remaining_e'} (the remaining
charges, electrons, in the trap)}

\item{\dots}{further arguments}

\item{r}{\link{numeric} (\emph{with default}): The radius of tunneling (unitless).}
}
\value{
This function returns an \code{\link{array}} with dimension length(times) x length(r) x clusters
}
\description{
Runs a Monte-Carlo (MC) simulation of thermo-luminesence (TL) using the tunneling (TUN) model. Tunneling refers to the direct movement of electrons from a trap directly to the recombination center
}
\details{
\deqn{
p(t) = s * e ^ (-E / kB * T) * e ^ (-r / rho ^ 1 / 3)
}
\deqn{
I_{TUN}(t) = 3 * n * p(t) *  r ^ 2 * e ^ (-r ^ 3) 
}

Where in the function \code{n} := \code{n_filled} := \code{t}:= \code{times}:= \eqn{\rho} := \eqn{\rho'} := \code{r_c} := \eqn{\rho_{c}'}
}
\section{Function version}{
 0.1.0
}

\examples{
\dontrun{
##============================================================================##
## Example 1: Simulate TL measurement
##============================================================================##
run_MC_TL_TUN(s = 3.5e12,
          E = 1.45,
          rho = 0.015,
          r_c = 0.85,
          times = 200:500) \%>\%
    plot_RLumCarlo(legend = T)
}

} 

\section{How to cite}{
Friedrich, J., Kreutzer, S., 2019. run_MC_TL_TUN(): Run Monte-Carlo Simulation for TL for tunnelling transition. Function version 0.1.0. In: Friedrich, J., Kreutzer, S., Pagonis, V., Schmidt, C., 2019. RLumCarlo: Monte-Carlo Methods for Simulating Luminescence PhenomenaR package version 0.1.0.9000-87. 
}

\references{
Pagonis, V. and Kulp, C., 2017. Monte Carlo simulations of tunneling phenomena and nearest neighbor hopping mechanism in feldspars. Journal of Luminescence 181, 114–120. \doi{10.1016/j.jlumin.2016.09.014}

Pagonis, V., Friedrich, J., Discher, M., Müller-Kirschbaum, A., Schlosser, V., Kreutzer, S., Chen, R. and Schmidt, C., 2019. Excited state luminescence signals from a random distribution of defects: A new Monte Carlo simulation approach for feldspar. Journal of Luminescence 207, 266–272. \doi{10.1016/j.jlumin.2018.11.024}

\strong{Further reading}
Aitken, M.J., 1985. Thermoluminescence dating. 276-280. \doi{10.1002/gea.3340020110}
}
\author{
Johannes Friedrich, University of Bayreuth (Germany), Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, Université Bordeaux Montaigne (France)
}
